requirements.txt
# requirements.txt
fish
ipykernel
nano
vim
https://github.com/glencoesoftware/zeroc-ice-py-linux-x86_64/releases/download/20240202/zeroc_ice-3.6.5-cp311-cp311-manylinux_2_28_x86_64.whl
https://github.com/laviolette-lab/lavlab-python-utils/releases/download/v1.3.1/lavlab_python_utils-1.3.1-py3-none-any.whl
omero-py
scikit-image==0.19.3
shapely
matplotlib
ipynb
tqdm
rasterio
valis-wsi==1.1.0
pillow==9.4.0
numpy==1.26
ipywidgets
